Abort this branch but take context with you
The existing path got too unwieldy so I changed gears and implemented
the move to a pure Dockerfile solution first to make it easier to scan
unpushed images. That also had other benefits like being able to pin
versions of tools.

I will cherry-pick this command on top of that work on main
to complete the Trivy scanning, using the lessons learned here
encoded in the plan.