Initialize the conatiner starter repo
This pattern will allow us to develop the starter in the
XPai repo and then push changes to the external repo
when ready.

Making it an external starter repo allows team to more
easily see what they need to integrate.

It also gives us a playground to make sure things work.