Just let projects configure mcp servers
For some reason, devcontainers don't complete the
auto registration of mcp servers in the init script.

it also might be better to just let projects configure the
MCP servers they want and we just make them available on the
image.

This also might be less of a problem as we spin up the
starting point repo for project.